+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FSL/helper/systems/MecanumSet.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.FSL.helper;
+package org.firstinspires.ftc.teamcode.FSL.helper.systems;
 
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Gamepad;
